+++ app/auth/signup/page.tsx
@@ -8,29 +8,20 @@ import { useForm } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { z } from 'zod';
 import Link from 'next/link';
-import { useMutation } from '@tanstack/react-query';
-import { useRouter } from 'next/navigation';
 import { LoaderCircle } from 'lucide-react';
-
-// todo
-// 1. zod 검사 파일
-// 2. mutation 로직 파일
-// 3. interface 파일
+import { SignupFormValues } from '@/interfaces/auth.interface';
+import { useSignupMutation } from '@/queries/auth/mutation';
 
 export default function SignupPage() {
-  const router = useRouter();
-  interface SignupForm {
-    email: string;
-    password: string;
-    name?: string;
-  }
+  const { mutate: signupMutation, isPending } = useSignupMutation();
+
   const registerSchema = z.object({
     email: z.string().min(1, { message: '이메일을 입력해주세요.' }).email('이메일 형식이 올바르지 않습니다.'),
     password: z.string().min(8, { message: '8자 이상 입력해주세요.' }),
     name: z.string(),
   });
 
-  const form = useForm<SignupForm>({
+  const form = useForm<SignupFormValues>({
     resolver: zodResolver(registerSchema),
     defaultValues: {
       email: '',
@@ -39,31 +30,10 @@ export default function SignupPage() {
     },
   });
 
-  const onSubmit = (data: SignupForm) => {
-    signup(data);
+  const onSubmit = (data: SignupFormValues) => {
+    signupMutation(data);
   };
 
-  const { mutate: signup, isPending } = useMutation<SignupForm, Error, SignupForm>({
-    mutationKey: ['signup'],
-    mutationFn: async (data: SignupForm) => {
-      const formData = new FormData();
-      formData.append('email', data.email);
-      formData.append('password', data.password);
-      if (data.name) {
-        formData.append('name', data.name);
-      }
-
-      const response = await fetch('/api/auth/signup', {
-        method: 'POST',
-        body: formData,
-      });
-
-      if (response.url) {
-        router.push(response.url);
-      }
-      return data;
-    },
-  });
   return (
     <div className="flex justify-center items-center h-screen ">
       <Card className="w-full max-w-md p-10">
