+++ app/getAllList/page.tsx
@@ -1,63 +1,24 @@
 'use client';
 
-import { useMutation, useQuery } from '@tanstack/react-query';
 import { Accordion, AccordionContent, AccordionItem, AccordionTrigger } from '@/components/ui/accordion';
 import { Checkbox } from '@/components/ui/checkbox';
-import { useQueryClient } from '@tanstack/react-query';
-
-interface TodoProps {
-  id: number;
-  title: string;
-  description: string;
-  completed: boolean;
-}
-
-const fetchAllList = async () => {
-  const response = await fetch('/api/todos', {
-    method: 'GET',
-  });
-
-  if (!response.ok) {
-    throw new Error('네트워크 에러');
-  }
-
-  const data = await response.json();
-  console.log(data.todos);
-  return data.todos;
-};
-
-const patchCompletedList = async (id: number, completed: boolean) => {
-  const response = await fetch(`/api/todos/${id}`, {
-    method: 'PATCH',
-    headers: {
-      'Content-Type': 'application/json',
-    },
-    body: JSON.stringify({ completed: !completed }),
-  });
-  const data = await response.json();
-  return data;
-};
+import { LoaderCircle } from 'lucide-react';
+import { TodoProps } from '@/interfaces/todos.interface';
+import { useGetAllList } from '@/queries/todos/queries';
+import { usePatchCompletedList } from '@/queries/todos/mutation';
 
 export default function GetAllList() {
-  const { data, isLoading, error } = useQuery({
-    queryKey: ['todoAllList'],
-    queryFn: fetchAllList,
-  });
-
-  const queryClient = useQueryClient();
-
-  const { mutate: patchCompleted } = useMutation({
-    mutationFn: ({ id, completed }: { id: number; completed: boolean }) => patchCompletedList(id, completed),
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ['todoAllList'] });
-    },
-  });
+  const { data, isLoading, error } = useGetAllList();
+  const { mutate: patchCompleted } = usePatchCompletedList();
 
   return (
     <div>
       <h1>투두 리스트</h1>
       {isLoading ? (
-        <div>로딩 중...</div>
+        <div className="flex justify-center items-center h-screen">
+          <LoaderCircle className="mr-2 h-4 w-4 animate-spin" />
+          로딩 중...
+        </div>
       ) : error ? (
         <div>에러가 발생했습니다: {error.message}</div>
       ) : (
@@ -70,7 +31,7 @@ export default function GetAllList() {
                     <Checkbox
                       checked={todo.completed}
                       className="flex mt-5 mr-5"
-                      onClick={() => patchCompleted({ id: todo.id, completed: todo.completed })}
+                      onCheckedChange={() => patchCompleted({ id: todo.id, completed: !todo.completed })}
                     />
                     <div className="flex-1">
                       <AccordionTrigger>
