+++ interfaces/auth.interface.ts
@@ -1,0 +1,10 @@
+export interface LoginFormValues {
+  email: string;
+  password: string;
+}
+
+export interface SignupFormValues {
+  email: string;
+  password: string;
+  name: string;
+}
