+++ interfaces/todos.interface.ts
@@ -1,0 +1,6 @@
+export interface TodoProps {
+  id: number;
+  title: string;
+  description: string;
+  completed: boolean;
+}
